fs/devfs: Use same function for idesc open
